# specs/008-api-v3-departamentos/contracts/empleados-v3.openapi.yaml
openapi: 3.0.3
info:
  title: API de Empleados y Departamentos v3
  version: 3.0.0
  description: >-
    Contrato publico v3 para gestion de departamentos y relacion de empleados
    por departamento. Incluye CRUD de departamentos, endpoint relacional
    obligatorio y extension de empleados para asignacion opcional de
    departamento con validaciones de integridad.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
tags:
  - name: DepartamentosV3
    description: CRUD y consultas relacionales de departamentos
  - name: EmpleadosV3
    description: Endpoints de empleados en contrato v3

paths:
  /api/v3/departamentos:
    get:
      tags: [DepartamentosV3]
      summary: Listar departamentos
      operationId: listarDepartamentosV3
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Lista paginada de departamentos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartamentoPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [DepartamentosV3]
      summary: Crear departamento
      operationId: crearDepartamentoV3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartamentoCreateRequest'
      responses:
        '201':
          description: Departamento creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartamentoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v3/departamentos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: '^DEP-[0-9]{6}$'
    get:
      tags: [DepartamentosV3]
      summary: Obtener departamento por ID
      operationId: obtenerDepartamentoV3
      responses:
        '200':
          description: Departamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartamentoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [DepartamentosV3]
      summary: Actualizar departamento
      operationId: actualizarDepartamentoV3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartamentoUpdateRequest'
      responses:
        '200':
          description: Departamento actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartamentoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [DepartamentosV3]
      summary: Eliminar departamento
      operationId: eliminarDepartamentoV3
      responses:
        '204':
          description: Departamento eliminado
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /api/v3/departamentos/{id}/empleados:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          pattern: '^DEP-[0-9]{6}$'
    get:
      tags: [DepartamentosV3]
      summary: Listar empleados por departamento
      operationId: listarEmpleadosPorDepartamentoV3
      description: Endpoint relacional obligatorio del contrato publico v3.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Lista paginada de empleados del departamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v3/empleados:
    post:
      tags: [EmpleadosV3]
      summary: Crear empleado (v3)
      operationId: crearEmpleadoV3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoCreateRequestV3'
      responses:
        '201':
          description: Empleado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponseV3'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

  /api/v3/empleados/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    put:
      tags: [EmpleadosV3]
      summary: Actualizar empleado (v3)
      operationId: actualizarEmpleadoV3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoUpdateRequestV3'
      responses:
        '200':
          description: Empleado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponseV3'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >-
        Autenticacion HTTP Basic para rutas protegidas.
        Username = correo_electronico, password = contrasena transitoria.

  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25

  schemas:
    DepartamentoCreateRequest:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 150
      required: [nombre]

    DepartamentoUpdateRequest:
      allOf:
        - $ref: '#/components/schemas/DepartamentoCreateRequest'

    DepartamentoResponse:
      type: object
      properties:
        id:
          type: string
          pattern: '^DEP-[0-9]{6}$'
          example: DEP-000001
        nombre:
          type: string
          maxLength: 150
      required: [id, nombre]

    EmpleadoCreateRequestV3:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        direccion:
          type: string
          minLength: 1
          maxLength: 100
        telefono:
          type: string
          minLength: 1
          maxLength: 100
        correoElectronico:
          type: string
          format: email
          maxLength: 150
        contrasena:
          type: string
          minLength: 8
          maxLength: 255
        departamentoId:
          type: string
          nullable: true
          pattern: '^DEP-[0-9]{6}$'
          description: Mapea al campo de persistencia `departamento_id`.
      required: [nombre, direccion, telefono, correoElectronico, contrasena]

    EmpleadoUpdateRequestV3:
      type: object
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        direccion:
          type: string
          minLength: 1
          maxLength: 100
        telefono:
          type: string
          minLength: 1
          maxLength: 100
        departamentoId:
          type: string
          nullable: true
          pattern: '^DEP-[0-9]{6}$'
          description: Mapea al campo de persistencia `departamento_id`.

    EmpleadoResponseV3:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        direccion:
          type: string
        telefono:
          type: string
        correoElectronico:
          type: string
          format: email
        departamentoId:
          type: string
          nullable: true
      required: [id, nombre, direccion, telefono, correoElectronico]

    EmpleadoPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EmpleadoResponseV3'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
      required: [content, page, size, totalElements, totalPages]

    DepartamentoPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/DepartamentoResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
      required: [content, page, size, totalElements, totalPages]

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [code, message]

  responses:
    BadRequest:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Credenciales invalidas
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Rol sin permisos para la operacion
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflicto de relacion (departamento con empleados asociados)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Referencia semantica invalida (`departamentoId` inexistente)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'